+++ Web3-eKYC-LIT/package.json
@@ -17,6 +17,7 @@
     "@rainbow-me/rainbowkit": "^2.2.6",
     "@tailwindcss/vite": "^4.1.10",
     "@tanstack/react-query": "^5.55.3",
+    "@web3-storage/w3up-client": "^17.3.0",
     "ethers": "^6.14.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
